squashed the front camera contrast bug!
there was a bug which was leading to really poor contrast on the front
camera.  Initially I thought it was something with the camera, but
noticied it was doing the same thing on my breadboard ONLY with the
second camera (the "front" camera).  well, this was that bug, and it's
dead now.