two small bugfixes in dualcameracapture.ino
two small bugfixes in dualcameracapture.ino.  one is using continue
instead of break in a while loop which processes incoming serial data
(could be USB or bluetooth).  the other is a removal of a dumb hack
which was actually working around a bug in the PC software
(glassescapture, previously called captureStreamDual).