ml_rawfeatures_knearest__replayhack
This is dirty as the night is long.  I was crunched for time when I
wrote this, and it still belongs in the repo even if it is dirty.

The ML algorithm programs should either save state for later replay (or
processing) or, alternatively, continue pipe data for another program to
attach to.