changed permissions, most of the repo was modified
changed file permissions and now most of the repo thinks it was changed.
so, I'm checking it in!  marking ALL files as not executable.  Some
files like ardueyectrl.py or seriallistener.py should actually be
executable.  remark them as such when the time comes that you need to do
so.