fixed some bugs (mainly in stonyman.v)
fixed some bugs (mainly in stonyman.v), but the interface between the
stonyman driver and the ADC driver is updated too.  the primary bug is
the reset bug in stonyman.v (but maybe primary is incorrect, as other
bugs were large too).  The stonyman driver is still getting stuck during
picture capture somewhere.