fixed FIFO logic in stonyman driver APB3 wrapper
there is also some general cleanup of that code.  This verilog works as
long as the application takes care to throttle itself.  There is no busy
flag in the stonyman driver to tell an application not to start a new
conversion (and attempting to do so will cause bad things to happen).