insight-serv uses mask for cam12
it is appended with _sf because it is for the smartfusion (it could be
different than computed with the teensy, but it's not likely to be THAT
different as long as the voltages are similar and the camera has the
same settings.

there are actually two masks, a primary and an alternate, something I
haven't done for other cameras.