ADC completes faster, added stonyman.v features
updated the ADC controller to trigger complete faster (as soon as the
correct number of bits have been read), while still not raising the CS
line or allowing a conversion any earlier.  This has made the overall
image capture quite a bit faster (from ~29 Hz up to ~ 32.5 Hz).

added features to the stonyman.v controller to allow for a different
input pixel size and output pixel size with minimal processing
(subtracting a fixed value and then mapping the bottom bits to the
output pixel).  It is functionally the same right now.