LKM driver multi-cam finish interrupts support
starting multi-cam support by supporting capture done interrupts for
every camera.  this means that each camera is expected to be part of its
own control group (simultaneous capture of multiple cameras will not be
supported in the driver for now).  each camera now has two GPIO
interrupts, a capture done interrupt and a high water interrupt.  there
is also a lot of cleanup in the GPIO defines.